Address sixth review of scheduler capacity.
The sixth automated review raised zero fix items -- the landing round.
This batch takes its one documentation item and five of its six
considers, all comments, docstrings, plan text and one test-assertion
tightening; no behaviour changes.

The documentation item scopes a claim the plan made too broadly: round
four's verdict that duplicate instance_location rows are "benign" is
true only at cluster scope, where the unclaimed fold is restricted to
nodes holding a capacity row. Per-claim used_* is deliberately
namespace-wide, so a duplicated placement counts an instance twice
there -- exactly the counter the closed-accounting fix does not
protect, and a quota in phase 4. The plan's phase-4 claim-accounting
list gains a third bullet (de-duplicate by instance uuid, or rely on
phase 3 eliminating stale placement rows) and the _RECONCILE_USAGE_SQL
comment block mirrors it next to the legacy-column under-count note.

The considers: baseobject.py imports NODE_ACTIVE_STATES instead of
carrying a third hardcoded copy of the set (fixing in passing its
comment's pre-existing claim that errored nodes are included when the
list excludes them), and node.py's comment stops instructing
maintainers to hand-sync the copy that no longer exists. The schema
tests now assert sa.BigInteger rather than sa.Integer -- BigInteger
subclasses Integer, so the old assertions passed for either type and
round four's deliberate overflow-avoiding widening was unpinned;
mutating a column back to Integer now fails the test.
_reconcile_fetch_demand's docstring records that its window comparison
spans hosts (refs.created is client-written) like the metrics
freshness check, with the negative-age clamp as the mitigation. The
live test module docstring states the --serial requirement its
database-scoped collation flip depends on. And the elected loop's
comment now names _run_due_scheduled_jobs() instead of the
run_pending() call it replaced.

The BOUNDED_QUERY_TIMEOUT consider is recorded rather than coded: its
skipped-pass-is-harmless justification is release-scoped, so the plan
now requires phase 3 to pair enabling the counter guard with a
persistent-reconcile-failure decision.

Prompt: Mikal asked me to rebase the branch and repush after the
GitHub Actions outage swallowed the previous push's events, then to
commit the documentation batch I had triaged from the sixth review
round. No further re-review is requested: the round contained no fix
items and this batch changes no behaviour.

+++ shakenfist/tests/test_mariadb_capacity_schema.py
@@ -111,10 +111,13 @@ class CapacityTableDefinitionTestCase(base.ShakenFistTestCase):
         self.assertTrue(table.c['node_uuid'].primary_key)
 
         # expected_demand is a DOUBLE with a server-side default of zero;
-        # the counters are integers defaulting to zero.
+        # the counters are BIGINTs defaulting to zero. BigInteger is asserted
+        # rather than Integer (which BigInteger would also satisfy) so a
+        # future narrowing of the deliberate overflow-avoiding widening
+        # fails a test rather than requiring a migration.
         self.assertIsInstance(table.c['expected_demand'].type, sa.Double)
         for name in ('used_cpus', 'used_memory_mb', 'used_disk_gb'):
-            self.assertIsInstance(table.c[name].type, sa.Integer)
+            self.assertIsInstance(table.c[name].type, sa.BigInteger)
             self.assertIsNotNone(table.c[name].server_default,
                                  f'{name} should default to zero')
 
@@ -156,10 +159,12 @@ class CapacityTableDefinitionTestCase(base.ShakenFistTestCase):
         self.assertTrue(table.c['id'].primary_key)
         self.assertFalse(table.c['id'].autoincrement)
 
+        # BigInteger asserted rather than Integer for the same reason as the
+        # node capacity counters above: pin the deliberate widening.
         for name in EXPECTED_COLUMNS['cluster_capacity']:
             if name in ('id', 'updated_at'):
                 continue
-            self.assertIsInstance(table.c[name].type, sa.Integer)
+            self.assertIsInstance(table.c[name].type, sa.BigInteger)
 
         self.assertIsInstance(table.c['updated_at'].type, sa.DateTime)
